refactor: domain이 service에 의존하지 않도록 수정
